pyproject.toml
[tool.poetry]
name = "flatten-image-dataset"
version = "0.1.0"
description = ""
authors = ["Josh Murr <joshuadavidmurr@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
jupyter = "^1.0.0"
numpy = "^1.22.3"
keras = "^2.8.0"
tensorflow = "^2.8.0"
matplotlib = "^3.5.1"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
